Module Melange_mini_stdlib

Since others depend on this file, its public mli files **should not export types** introduced here, otherwise it would cause conflicts here.

If the type exported here is also exported in modules from others, you will get a type not equivalent.

Types defined here but should not export:

ref (make sure not exported in public others/*.mli)

val (^) : string -> string -> string

val (=) : 'a -> 'a -> bool

val (<>) : 'a -> 'a -> bool

val (==) : 'a -> 'a -> bool

val (!=) : 'a -> 'a -> bool

val (<) : 'a -> 'a -> bool

val (>) : 'a -> 'a -> bool

val (<=) : 'a -> 'a -> bool

val (>=) : 'a -> 'a -> bool

val (+) : int -> int -> int

val (-) : int -> int -> int

val (~-) : int -> int

val (*) : int -> int -> int

val (/) : int -> int -> int

val (lsl) : int -> int -> int

val (lor) : int -> int -> int

val (land) : int -> int -> int

val (mod) : int -> int -> int

val (lsr) : int -> int -> int

val (lxor) : int -> int -> int

val (asr) : int -> int -> int

type 'a ref = {

mutable contents : 'a;

}

val ref : 'a -> 'a ref

val (||) : bool -> bool -> bool

val (&&) : bool -> bool -> bool

val not : bool -> bool

val raise : exn -> 'a

val ignore : 'a -> unit

val (|>) : 'a -> ('a -> 'b) -> 'b

val (@@) : ('a -> 'b) -> 'a -> 'b

val (**) : float -> float -> float

val (~-.) : float -> float

val (+.) : float -> float -> float

val (-.) : float -> float -> float

val (*.) : float -> float -> float

val (/.) : float -> float -> float

module Obj : sig ... end

module Stdlib : sig ... end